rename imports in service/controller/kvm/v7patch1: github.com/giantswarm/cluster-operator/pkg/v7/ > github.com/giantswarm/cluster-operator/pkg/v7patch1/
