+++ service/controller/kvm/v7patch1/resource/configmap/resource.go
@@ -5,7 +5,7 @@ import (
 	"github.com/giantswarm/micrologger"
 	corev1 "k8s.io/api/core/v1"
 
-	"github.com/giantswarm/cluster-operator/pkg/v7/configmap"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/configmap"
 )
 
 const (
