+++ service/controller/kvm/v7patch1/resource_set.go
@@ -20,13 +20,13 @@ import (
 
 	"github.com/giantswarm/cluster-operator/pkg/cluster"
 	"github.com/giantswarm/cluster-operator/pkg/label"
-	chartconfigservice "github.com/giantswarm/cluster-operator/pkg/v7/chartconfig"
-	configmapservice "github.com/giantswarm/cluster-operator/pkg/v7/configmap"
-	"github.com/giantswarm/cluster-operator/pkg/v7/resource/certconfig"
-	"github.com/giantswarm/cluster-operator/pkg/v7/resource/chart"
-	"github.com/giantswarm/cluster-operator/pkg/v7/resource/clustercr"
-	"github.com/giantswarm/cluster-operator/pkg/v7/resource/encryptionkey"
-	"github.com/giantswarm/cluster-operator/pkg/v7/resource/namespace"
+	chartconfigservice "github.com/giantswarm/cluster-operator/pkg/v7patch1/chartconfig"
+	configmapservice "github.com/giantswarm/cluster-operator/pkg/v7patch1/configmap"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/resource/certconfig"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/resource/chart"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/resource/clustercr"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/resource/encryptionkey"
+	"github.com/giantswarm/cluster-operator/pkg/v7patch1/resource/namespace"
 	"github.com/giantswarm/cluster-operator/service/controller/kvm/v7patch1/key"
 	"github.com/giantswarm/cluster-operator/service/controller/kvm/v7patch1/resource/chartconfig"
 	"github.com/giantswarm/cluster-operator/service/controller/kvm/v7patch1/resource/configmap"
